feat: products CRD

+++ src/routes/api/files/+server.ts
@@ -1,7 +1,7 @@
 import { ApiResponse, errorHandler } from "$lib/server/apiResponse";
-import cloudinary from "$lib/server/cloudinary";
 import type { RequestEvent } from "@sveltejs/kit";
-import { deleteSchema, uploadSchema } from "./dtos";
+import { uploadSchema } from "./dtos";
+import { deleteFile, uploadFile } from "./utils";
 
 export async function POST(event: RequestEvent) {
     try {
@@ -15,7 +15,8 @@ export async function POST(event: RequestEvent) {
 }
 export async function DELETE(event: RequestEvent) {
     try {
-        const { file_url } = deleteSchema.parse(await event.request.json())
+        const file_url = new URL(event.request.url).searchParams.get("fileurl")
+        if (!file_url) throw new Error("No file url provided")
         const file = await deleteFile(file_url)
         return ApiResponse(200, true, 'File deleted', file)
     } catch (error) {
@@ -24,19 +25,3 @@ export async function DELETE(event: RequestEvent) {
 
 }
 
-async function uploadFile(file: string, filename: string) {
-    const res = await cloudinary.uploader.upload(file, {
-        access_mode: 'public',
-        public_id: filename
-            .replaceAll(/[. ]/g, '-')
-            .toLocaleLowerCase(),
-        allowed_formats: ['jpg', 'png', 'jpeg', 'webm', 'webp', 'pdf'],
-    });
-    return res.secure_url
-}
-
-async function deleteFile(url: string) {
-    const publicId = url.split('/').pop()!.split('.')[0];
-    const res = await cloudinary.uploader.destroy(publicId);
-    return res.result
-}
